+++ inst/doc/ggdemetra.R
@@ -3,7 +3,8 @@ knitr::opts_chunk$set(
   collapse = TRUE,
   comment = "#>",
   fig.dim = c(7,4)*1.4,
-  out.width = "100%"
+  out.width = "100%",
+  warning = FALSE
 )
 
 ## ---- eval = FALSE------------------------------------------------------------
@@ -25,15 +26,15 @@ sa <- jx13(ipi_c_eu[, "FR"])
 ## -----------------------------------------------------------------------------
 spec <- RJDemetra::x13_spec("RSA3", tradingdays.option = "WorkingDays")
 p_ipi_fr +
-    geom_sa(color = "red",
+    geom_sa(color = "#155692",
             spec = spec)
 
 ## -----------------------------------------------------------------------------
 p_sa <- p_ipi_fr +
     geom_sa(component = "y_f", linetype = 2, message = FALSE,
             spec = spec) + 
-    geom_sa(component = "sa", color = "red") +
-    geom_sa(component = "sa_f", color = "red", linetype = 2)
+    geom_sa(component = "sa", color = "#155692") +
+    geom_sa(component = "sa_f", color = "#155692", linetype = 2)
 p_sa
 
 ## -----------------------------------------------------------------------------
@@ -42,8 +43,7 @@ p_sa + geom_outlier(geom = "label")
 ## -----------------------------------------------------------------------------
 p_sa + 
     geom_outlier(geom = "label_repel",
-                 vjust = 4,
-                 ylim = c(NA, 65), force = 10,
+                 ylim = c(NA, 65), 
                  arrow = arrow(length = unit(0.03, "npc"),
                                type = "closed", ends = "last"))
 
@@ -51,8 +51,7 @@ p_sa +
 p_sa + 
     geom_outlier(geom = "label_repel",
                  first_date = 2009,
-                 vjust = 4,
-                 ylim = c(NA, 65), force = 10,
+                 ylim = c(NA, 65), 
                  arrow = arrow(length = unit(0.03, "npc"),
                                type = "closed", ends = "last"))
 
@@ -81,11 +80,43 @@ p_sa +
 ## -----------------------------------------------------------------------------
 p_diag <- ggplot(data = ipi_c_eu_df, mapping = aes(x = date, y = FR))  +
     geom_diagnostics(diagnostics = diagnostics,
-                     spec = spec,
-                     table_theme = gridExtra::ttheme_default(base_size = 8),
-                     message = FALSE) + 
+                     spec = spec, frequency = 12,
+                     table_theme = gridExtra::ttheme_default(base_size = 8)) + 
     theme_void()
     
 gridExtra::grid.arrange(p_sa, p_diag,
              nrow = 2, heights  = c(4, 1.5))
+
+## ----mod----------------------------------------------------------------------
+mod <- RJDemetra::x13(ipi_c_eu[,"UK"], spec)
+
+## ----init-ggplot--------------------------------------------------------------
+init_ggplot(mod) + 
+    geom_line(color =  "#F0B400") +
+    geom_sa(color =  "#155692") +
+    geom_arima(geom = "label",
+               x_arima = -Inf, y_arima = -Inf, 
+               vjust = -1, hjust = -0.1)
+
+## ----sa-init------------------------------------------------------------------
+data <- ts.union(raw(mod), raw(mod, forecast = TRUE),
+                 trendcycle(mod), trendcycle(mod, forecast = TRUE),
+                 seasonaladj(mod), seasonaladj(mod, forecast = TRUE))
+colnames(data) <- c("y", "y_f",
+                    "t", "t_f",
+                    "sa", "sa_f")
+ggplot(data = ts2df(data), mapping = aes(x = date)) +
+    geom_line(mapping = aes(y = y), color =  "#F0B400", na.rm = TRUE) +
+    geom_line(mapping = aes(y = y_f), color =  "#F0B400", na.rm = TRUE, linetype = 2) +
+    geom_line(mapping = aes(y = t), color =  "#1E6C0B", na.rm = TRUE) +
+    geom_line(mapping = aes(y = t_f), color =  "#1E6C0B", na.rm = TRUE, linetype = 2) +
+    geom_line(mapping = aes(y = sa), color =  "#155692", na.rm = TRUE) +
+    geom_line(mapping = aes(y = sa_f), color =  "#155692", na.rm = TRUE, linetype = 2) +
+    theme_bw()
+
+## ----ggsiratio----------------------------------------------------------------
+ggsiratioplot(mod)
+
+## ----autoplot-----------------------------------------------------------------
+autoplot(mod)