+++ NEWS.md
@@ -1,4 +1,12 @@
-# ggdemetra 0.2.4
+# ggdemetra 0.2.6
+
+* `siratio()` correction with TRAMO-SEATS `"jSA"` models.
+
+* `y_forecast()` replaced by `raw()`.
+
+* new `init_ggplot()` function.
+
+# ggdemetra 0.2.5
 
 * new functions: `siratio()`, `siratioplot()` and `ggsiratioplot()` to plot SI ratios.